refactor: introduce  interface for RandomNumberGenerator abstraction - update Car member varriable and method to use abstact type
