+++ README.md
@@ -30,6 +30,8 @@
     - Throw an `IllegalArgumentException` if the value is out of range
 
 - [x] Separate car names using comma
-- [x] Generate a random number between 0 and 9 (use `Randoms.pickNumberInRange()`)
+- [x] Generate a number between 0 and 9
+    - [x] Generate a random number between 0 and 9 (use `Randoms.pickNumberInRange()`)
+    - [x] Generate a fixed number between 0 and 9
 - [x] Move a car forward only if the randomly generated number is 4 or greater
 - [x] draw the winner
