+++ src/main/kotlin/racingcar/domain/numbergenerator/NumberGenerator.kt
@@ -1,0 +1,6 @@
+package racingcar.domain.numbergenerator
+
+interface NumberGenerator {
+
+    fun generate(): Int
+}
